+++ app/assets/javascripts/models/note.js
@@ -12,7 +12,7 @@ var Note = module.exports = Backbone.Model.extend({
     if( model.isReadonly ) {
       return true;
     }
-    Backbone.sync(method, model, options);
+    return Backbone.sync(method, model, options);
   }
 
 });